(fix)音源の音質周りのバグ修正
内容	ファイル
FM フィードバックを実機準拠（エンベロープ適用後）に	オペレータ 6種
FM の位相を正規化 + double に	FmOperator + オペレータ 6種
SSG / BEEP の位相折り返しを while に	SynthSsg.cpp / SynthBeep.cpp
BEEP のエイリアス抑制（PolyBLEP）と OPTIONAL カテゴリ	BEEP の DSP / パラメータ / GUI
OPL/OPL3 のエンベロープ二重適用の修正	EnvOplAdsr.cpp

+++ 2686VLight/Source/Core/Fm/FmOperator.h
@@ -38,8 +38,10 @@ protected:
     enum class State { Idle, Attack, Decay, Sustain, Release };
     State m_state = State::Idle;
     double m_sampleRate = 44100.0;
-    float m_phase = 0.0f;
-    float m_phaseDelta = 0.0f;
+    // 位相はサイクル単位 (1.0 で 1 周)。
+    // 高域でも位相のずれが溜まりにくいよう double で持つ。
+    double m_phase = 0.0;
+    double m_phaseDelta = 0.0;
     double m_ssgPhase = 0.0;
     float m_ssgEgFreq = 1.0f;
     int m_noteNumber = 60;
